+++ lib/center_extension.dart
@@ -1,0 +1,9 @@
+part of 'awesome_extensions.dart';
+
+extension CenterExtension on Widget {
+  Center toCenter() {
+    return Center(
+      child: this,
+    );
+  }
+}